nightly: remove some detritus
Silence a lint warning about usage of `testing.T.Fatal`.

Release note: None
